Move Scarb and CLI config combining to main
