+++ crates/forge/src/main.rs
@@ -1,5 +1,5 @@
 use anyhow::{anyhow, bail, Context, Result};
-use camino::Utf8PathBuf;
+use camino::{Utf8Path, Utf8PathBuf};
 use clap::{Parser, Subcommand, ValueEnum};
 use forge::scarb::config_from_scarb_for_package;
 use include_dir::{include_dir, Dir};
@@ -15,11 +15,15 @@ use std::{env, fs};
 use tempfile::{tempdir, TempDir};
 use tokio::runtime::Builder;
 
-use forge::{pretty_printing, CancellationTokens, RunnerConfig, RunnerParams, CACHE_DIR};
+use forge::{
+    pretty_printing, CancellationTokens, RunnerConfig, RunnerParams, CACHE_DIR, FUZZER_RUNS_DEFAULT,
+};
 use forge::{run, TestCrateSummary};
 
+use forge::scarb::config::ForgeConfig;
 use forge::test_case_summary::TestCaseSummary;
 use forge::test_filter::TestsFilter;
+use rand::{thread_rng, RngCore};
 use std::process::{Command, Stdio};
 use std::thread::available_parallelism;
 
@@ -129,6 +133,26 @@ fn extract_failed_tests(tests_summaries: Vec<TestCrateSummary>) -> Vec<TestCaseS
         .collect()
 }
 
+fn combine_configs(
+    workspace_root: &Utf8Path,
+    exit_first: bool,
+    fuzzer_runs: Option<u32>,
+    fuzzer_seed: Option<u64>,
+    forge_config: &ForgeConfig,
+) -> RunnerConfig {
+    RunnerConfig::new(
+        workspace_root.to_path_buf(),
+        exit_first || forge_config.exit_first,
+        forge_config.fork.clone(),
+        fuzzer_runs
+            .or(forge_config.fuzzer_runs)
+            .unwrap_or(FUZZER_RUNS_DEFAULT),
+        fuzzer_seed
+            .or(forge_config.fuzzer_seed)
+            .unwrap_or_else(|| thread_rng().next_u64()),
+    )
+}
+
 fn test_workspace(args: TestArgs) -> Result<bool> {
     match args.color {
         ColorOption::Always => env::set_var("CLICOLOR_FORCE", "1"),
@@ -187,8 +211,8 @@ fn test_workspace(args: TestArgs) -> Result<bool> {
 
                 let contracts = get_contracts_map(&scarb_metadata, &package.id).unwrap_or_default();
 
-                let runner_config = Arc::new(RunnerConfig::new(
-                    workspace_root.clone(),
+                let runner_config = Arc::new(combine_configs(
+                    &workspace_root,
                     args.exit_first,
                     args.fuzzer_runs,
                     args.fuzzer_seed,
@@ -270,4 +294,88 @@ fn main() {
             std::process::exit(2);
         }
     };
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn fuzzer_default_seed() {
+        let workspace_root: Utf8PathBuf = Default::default();
+        let config = combine_configs(&workspace_root, false, None, None, &Default::default());
+        let config2 = combine_configs(&workspace_root, false, None, None, &Default::default());
+
+        assert_ne!(config.fuzzer_seed, 0);
+        assert_ne!(config2.fuzzer_seed, 0);
+        assert_ne!(config.fuzzer_seed, config2.fuzzer_seed);
+    }
+
+    #[test]
+    fn runner_config_default_arguments() {
+        let workspace_root: Utf8PathBuf = Default::default();
+        let config = combine_configs(&workspace_root, false, None, None, &Default::default());
+        assert_eq!(
+            config,
+            RunnerConfig {
+                workspace_root,
+                exit_first: false,
+                fork_targets: vec![],
+                fuzzer_runs: FUZZER_RUNS_DEFAULT,
+                fuzzer_seed: config.fuzzer_seed,
+            }
+        );
+    }
+
+    #[test]
+    fn runner_config_just_scarb_arguments() {
+        let config_from_scarb = ForgeConfig {
+            exit_first: true,
+            fork: vec![],
+            fuzzer_runs: Some(1234),
+            fuzzer_seed: Some(500),
+        };
+        let workspace_root: Utf8PathBuf = Default::default();
+
+        let config = combine_configs(&workspace_root, false, None, None, &config_from_scarb);
+        assert_eq!(
+            config,
+            RunnerConfig {
+                workspace_root,
+                exit_first: true,
+                fork_targets: vec![],
+                fuzzer_runs: 1234,
+                fuzzer_seed: 500,
+            }
+        );
+    }
+
+    #[test]
+    fn runner_config_argument_precedence() {
+        let workspace_root: Utf8PathBuf = Default::default();
+
+        let config_from_scarb = ForgeConfig {
+            exit_first: false,
+            fork: vec![],
+            fuzzer_runs: Some(1234),
+            fuzzer_seed: Some(1000),
+        };
+        let config = combine_configs(
+            &workspace_root,
+            true,
+            Some(100),
+            Some(32),
+            &config_from_scarb,
+        );
+        assert_eq!(
+            config,
+            RunnerConfig {
+                workspace_root,
+                exit_first: true,
+                fork_targets: vec![],
+                fuzzer_runs: 100,
+                fuzzer_seed: 32,
+            }
+        );
+    }
 }
