+++ crates/forge/test_utils/src/running_tests.rs
@@ -6,7 +6,6 @@ use camino::Utf8PathBuf;
 
 use forge::test_filter::TestsFilter;
 use forge::{run, CancellationTokens, RunnerConfig, RunnerParams, TestCrateSummary};
-use std::default::Default;
 use std::path::PathBuf;
 use tempfile::tempdir;
 use tokio::runtime::Runtime;
@@ -23,9 +22,9 @@ pub fn run_test_case(test: &TestCase) -> Vec<TestCrateSummary> {
         Arc::new(RunnerConfig::new(
             Utf8PathBuf::from_path_buf(PathBuf::from(tempdir().unwrap().path())).unwrap(),
             false,
-            Some(256),
-            Some(12345),
-            &Default::default(),
+            vec![],
+            256,
+            12345,
         )),
         Arc::new(RunnerParams::new(
             corelib_path(),
